{{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/requirements/common.txt
pip>=21.2.2
concurrent-log-handler==0.9.24
cryptography==38.0.1
setuptools==57.1.0
psycopg2==2.9.5
python-dotenv==0.21.0
Django==4.2.4
django-ckeditor==6.7.0
django-compressor==4.4
django-constance==2.9.1
django-picklefield==3.1
django-cleanup==8.0.0
django-extensions==3.2.3
django-getenv==1.3.2
django-taggit==4.0.0
django-browser-reload==1.11.0
django-tailwind==3.6.0
django-libsass==0.9
django-user-agents==0.4.0
django-mptt==0.14.0
django-lineup==0.3.1
{% if cookiecutter.admin == 'django-baton' %}
django-baton==2.7.1
{% elif cookiecutter.admin == 'django-grappelli' %}
django-grappelli==3.0.3
{% endif %}
{% if cookiecutter.use_disqus == 'y' %}
django-disqus==0.5
{% endif %}
{% if cookiecutter.use_filer == 'y' %}
django-filer==2.2.3
easy-thumbnails==2.8.3
{% endif %}
{% if cookiecutter.use_subject_imagefield == 'y' %}
django-subject-imagefield==0.2.12
{% endif %}
{% if cookiecutter.use_sorl_thumbnail == 'y' %}
sorl-thumbnail==12.9.0
{% endif %}
{% if cookiecutter.use_simple_captcha == 'y' %}
django-simple-captcha==0.5.18
{% endif %}
